refactor FAQ
